Change SearchBar Styling and Refactor
Moved SearchBar to components and created new SearchQuestions page
